+++ src/uicomponents/CMakeLists.txt
@@ -11,9 +11,8 @@ set(MODULE_QML_IMPORT ${CMAKE_CURRENT_LIST_DIR}/qml)
 set(MODULE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/uicomponentsmodule.cpp
     ${CMAKE_CURRENT_LIST_DIR}/uicomponentsmodule.h
+    ${CMAKE_CURRENT_LIST_DIR}/types/numerictypes.h
 
-    ${CMAKE_CURRENT_LIST_DIR}/components/numericviewmodel.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/components/numericviewmodel.h
     ${CMAKE_CURRENT_LIST_DIR}/components/bpmmodel.cpp
     ${CMAKE_CURRENT_LIST_DIR}/components/bpmmodel.h
     ${CMAKE_CURRENT_LIST_DIR}/components/frequencymodel.cpp
@@ -22,15 +21,18 @@ set(MODULE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/components/timecodemodel.h
 
     ${CMAKE_CURRENT_LIST_DIR}/components/timecodemodeselector.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/components/timecodeformatter.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/components/timecodeformatter.h
-    ${CMAKE_CURRENT_LIST_DIR}/components/numericformatter.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/components/numericformatter.h
-    ${CMAKE_CURRENT_LIST_DIR}/components/beatsformatter.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/components/beatsformatter.h
-    ${CMAKE_CURRENT_LIST_DIR}/components/fieldsinteractioncontroller.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/components/fieldsinteractioncontroller.h
-    ${CMAKE_CURRENT_LIST_DIR}/types/numerictypes.h
+
+    ${CMAKE_CURRENT_LIST_DIR}/components/internal/numeric/numericviewmodel.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/components/internal/numeric/numericviewmodel.h
+    ${CMAKE_CURRENT_LIST_DIR}/components/internal/numeric/timecodeformatter.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/components/internal/numeric/timecodeformatter.h
+    ${CMAKE_CURRENT_LIST_DIR}/components/internal/numeric/numericformatter.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/components/internal/numeric/numericformatter.h
+    ${CMAKE_CURRENT_LIST_DIR}/components/internal/numeric/beatsformatter.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/components/internal/numeric/beatsformatter.h
+    ${CMAKE_CURRENT_LIST_DIR}/components/internal/numeric/fieldsinteractioncontroller.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/components/internal/numeric/fieldsinteractioncontroller.h
+
 )
 
 setup_module()
