+++ src/uicomponents/components/internal/numeric/numericformatter.cpp
